[add]添加数据导出模块，修复部分代码错误

+++ ad-service/ad-sponsor/src/main/java/com/glw/ad/dao/AdPlanRepository.java
@@ -19,5 +19,5 @@ public interface AdPlanRepository extends JpaRepository<AdPlan, Long> {
 
     AdPlan findByUserIdAndPlanName(Long userId, String planName);
 
-    List<AdPlan> findAllByPlanStatus(Integer status);
+    List<AdPlan> findAllByPlanStatus(Byte status);
 }
