+++ ad-service/ad-sponsor/src/main/java/com/glw/ad/dao/AdUnitRepository.java
@@ -15,5 +15,5 @@ public interface AdUnitRepository extends JpaRepository<AdUnit, Long> {
 
     AdUnit findByPlanIdAndUnitName(Long planId, String unitName);
 
-    List<AdUnit> findAllByUnitStatus(Integer unitStatus);
+    List<AdUnit> findAllByUnitStatus(Byte unitStatus);
 }
